requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.1=py38h0a891b7_1
aiosignal=1.3.1=pyhd8ed1ab_0
appdirs=1.4.4=pyh9f0ad1d_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
backcall=0.2.0=pyh9f0ad1d_0
backoff=2.2.1=pyhd8ed1ab_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=39.0.1=py38h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.11.0=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
et_xmlfile=1.1.0=pyhd8ed1ab_0
evaluate=0.4.0=pypi_0
executing=1.2.0=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py38h06a4308_0
flit-core=3.8.0=py38h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py38h5eee18b_0
fsspec=2023.3.0=pypi_0
giflib=5.2.1=h5eee18b_3
git-lfs=3.3.0=ha770c72_0
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.31=pyhd8ed1ab_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface_hub=0.13.3=py_0
idna=3.4=py38h06a4308_0
importlib-metadata=6.0.0=py38h06a4308_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.13=pyhd8ed1ab_0
ipython=8.12.0=pyh41d4057_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py38h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.6.0.25=0
libcurand=10.3.2.56=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib-base=3.4.3=py38hf4fb855_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py38h06a4308_0
multidict=6.0.2=py38h5eee18b_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_1
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
openai=0.27.2=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.0.9=pyhd8ed1ab_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=py38h06a4308_0
pandas=1.5.2=py38h417a72b_0
pandas-stubs=1.5.3.230321=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=py_1
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h6a678d5_0
pip=23.0.1=py38h06a4308_0
plotly=5.14.0=pyhd8ed1ab_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=3.20.3=py38h6a678d5_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pyhd8ed1ab_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_3
pytorch-mutex=1.0=cuda
pytz=2022.7=py38h06a4308_0
pyyaml=6.0=py38h5eee18b_1
readline=8.2=h5eee18b_0
regex=2022.7.9=py38h5eee18b_0
requests=2.28.1=py38h06a4308_1
responses=0.18.0=pypi_0
sacremoses=master=py_0
scikit-learn=1.2.2=py38h6a678d5_0
scipy=1.9.3=py38h14f4228_1
sentencepiece=0.1.97=pypi_0
sentry-sdk=1.18.0=pyhd8ed1ab_0
setproctitle=1.2.2=py38h0a891b7_2
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.41.1=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py38h06a4308_0
tenacity=8.2.2=pyhd8ed1ab_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.0.dev0=py38_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=2.0.0=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu117
tornado=6.1=py38h0a891b7_3
tqdm=4.65.0=py38hb070fc8_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.28.0.dev0=pypi_0
types-pytz=2023.3.0.0=pyhd8ed1ab_0
typing-extensions=4.4.0=py38h06a4308_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.15=py38h06a4308_0
wandb=0.14.0=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=py38h0a891b7_2
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.4=hc292b87_0
